moved generate_prediction to main.py
