+++ utils.py
@@ -343,48 +343,3 @@ def coco_evaluation(gt_filepath, prediction_filepath, image_ids, category_ids):
     cocoEval.summarize()
 
     return cocoEval.stats[1], cocoEval.stats[8]
-
-
-def generate_prediction(meta_data, pipeline, logger, category_ids, chunk_size):
-    if chunk_size is not None:
-        return _generate_prediction_in_chunks(meta_data, pipeline, logger, category_ids, chunk_size)
-    else:
-        return _generate_prediction(meta_data, pipeline, logger, category_ids)
-
-
-def _generate_prediction(meta_data, pipeline, logger, category_ids):
-    data = {'input': {'meta': meta_data,
-                      'meta_valid': None,
-                      'train_mode': False,
-                      'target_sizes': [(300, 300)] * len(meta_data),
-                      },
-            }
-
-    pipeline.clean_cache()
-    output = pipeline.transform(data)
-    pipeline.clean_cache()
-    y_pred = output['y_pred']
-
-    prediction = create_annotations(meta_data, y_pred, logger, category_ids)
-    return prediction
-
-
-def _generate_prediction_in_chunks(meta_data, pipeline, logger, category_ids, chunk_size):
-    prediction = []
-    for meta_chunk in generate_data_frame_chunks(meta_data, chunk_size):
-        data = {'input': {'meta': meta_chunk,
-                          'meta_valid': None,
-                          'train_mode': False,
-                          'target_sizes': [(300, 300)] * len(meta_chunk)
-                          },
-                }
-
-        pipeline.clean_cache()
-        output = pipeline.transform(data)
-        pipeline.clean_cache()
-        y_pred = output['y_pred']
-
-        prediction_chunk = create_annotations(meta_chunk, y_pred, logger, category_ids)
-        prediction.extend(prediction_chunk)
-
-    return prediction
